requirements.txt
# Core packages
python-dotenv
numpy
Pillow
matplotlib
accelerate

# ML packages
diffusers
transformers
torch
torchvision
safetensors
omegaconf

# Gradio and Spaces
gradio
spaces

# API
fastapi
uvicorn
python-multipart
httpx
pydantic

modal
